pyproject.toml
[project]
name = "videosrt"
version = "0.1.0"
description = "MP4 to SRT subtitle generator with YouTube metadata and video editing"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "openai>=1.54.0",
    "google-genai>=1.0.0",
    "cryptography>=42.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
